Iconify button now iconifies entire switcher, not just that particular item.
Currently, iconify button on dock item is rather useless because one has to repeatedly
iconify all the items in a switcher one by one anyways to recover the screen real-estate.
Iconifying the item alone does not help anything since switcher tabs more or less does
the job of switching to different dock items for viewing purpose. Using iconify button
as a means of recovering the real-estate is more practical and useful.