bgo #569160 - Anjuta doesn't remember main window layout
GdlDockLayout creates hidden GdlDockItem without a child widget for named
widget not existing yet. When adding  a new widget GdlDock checks if
such hidden widget exists. If it is the case the hidden widget is replaced by the
new one, using the last saved position. If not the new widget is docked at its
default position like before. So a layout affects not only existing widgets but
widgets created afterward if they were existing before.