Allow to display widget not present in the saved layout
When you load a new layout, widgets present in the old layout and not in the
new one are hidden but it was not possible to re display them later because
their parents have been replaced by widgets from the saved layout.

Now, all widgets are removed from the layout and when you try to display
a widget not already present in the widget hierarchy, it is not just shown but
it is docked again as it was done before.