Consider iconified widget as closed
It allows to manage the flag GDL_DOCK_ATTACHED in the GtkWidget hide and show
functions. We don't need to hide a widget without clearing this flag. It means
that an iconified widget cannot be closed or not, it is always closed so there
are 3 states for a dock item: normal, iconified, closed (including detached).